requirements.txt
fastapi
uvicorn
python-whois
dnspython
redis
aioredis
validate-email-address
